client/requirements.txt
pycryptodome==3.19.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pymysql==1.1.0
